+++ packages/liveblocks-core/src/comments/index.ts
@@ -1,10 +1,19 @@
 import type { AuthValue } from "../auth-manager";
 import type { JsonObject } from "../lib/Json";
 import type { Polyfills } from "../room";
+import {
+  convertToCommentData,
+  convertToCommentUserReaction,
+  convertToThreadData,
+} from "./comment-body";
 import type { BaseMetadata } from "./types/BaseMetadata";
 import type { CommentBody } from "./types/CommentBody";
-import type { CommentData } from "./types/CommentData";
-import type { ThreadData } from "./types/ThreadData";
+import type { CommentData, CommentDataPlain } from "./types/CommentData";
+import type {
+  CommentUserReaction,
+  CommentUserReactionPlain,
+} from "./types/CommentReaction";
+import type { ThreadData, ThreadDataPlain } from "./types/ThreadData";
 
 function getAuthBearerHeaderFromAuthValue(authValue: AuthValue) {
   if (authValue.type === "public") {
@@ -34,7 +43,7 @@ export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
   editThreadMetadata(options: {
     metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
-  }): Promise<ThreadData<TThreadMetadata>>;
+  }): Promise<TThreadMetadata>;
   createComment(options: {
     threadId: string;
     commentId: string;
@@ -53,12 +62,12 @@ export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
     threadId: string;
     commentId: string;
     emoji: string;
-  }): Promise<CommentData>;
+  }): Promise<CommentUserReaction>;
   removeReaction(options: {
     threadId: string;
     commentId: string;
     emoji: string;
-  }): Promise<CommentData>;
+  }): Promise<void>;
 };
 
 export class CommentsApiError extends Error {
@@ -139,9 +148,9 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
 
     if (response.ok) {
       const json = await (response.json() as Promise<{
-        data: ThreadData<TThreadMetadata>[];
+        data: ThreadDataPlain<TThreadMetadata>[];
       }>);
-      return json.data;
+      return json.data.map((thread) => convertToThreadData(thread));
     } else if (response.status === 404) {
       return [];
     } else {
@@ -149,7 +158,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     }
   }
 
-  function createThread({
+  async function createThread({
     metadata,
     body,
     commentId,
@@ -161,23 +170,28 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     metadata: TThreadMetadata | undefined;
     body: CommentBody;
   }) {
-    return fetchJson<ThreadData<TThreadMetadata>>("/threads", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        id: threadId,
-        comment: {
-          id: commentId,
-          body,
+    const thread = await fetchJson<ThreadDataPlain<TThreadMetadata>>(
+      "/threads",
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
         },
-        metadata,
-      }),
-    });
+        body: JSON.stringify({
+          id: threadId,
+          comment: {
+            id: commentId,
+            body,
+          },
+          metadata,
+        }),
+      }
+    );
+
+    return convertToThreadData(thread);
   }
 
-  function editThreadMetadata({
+  async function editThreadMetadata({
     metadata,
     threadId,
   }: {
@@ -185,7 +199,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
   }) {
-    return fetchJson<ThreadData<TThreadMetadata>>(
+    return await fetchJson<TThreadMetadata>(
       `/threads/${encodeURIComponent(threadId)}/metadata`,
       {
         method: "POST",
@@ -197,7 +211,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     );
   }
 
-  function createComment({
+  async function createComment({
     threadId,
     commentId,
     body,
@@ -206,7 +220,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     body: CommentBody;
   }) {
-    return fetchJson<CommentData>(
+    const comment = await fetchJson<CommentDataPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments`,
       {
         method: "POST",
@@ -219,9 +233,11 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         }),
       }
     );
+
+    return convertToCommentData(comment);
   }
 
-  function editComment({
+  async function editComment({
     threadId,
     commentId,
     body,
@@ -230,7 +246,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     body: CommentBody;
   }) {
-    return fetchJson<CommentData>(
+    const comment = await fetchJson<CommentDataPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}`,
@@ -244,6 +260,8 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         }),
       }
     );
+
+    return convertToCommentData(comment);
   }
 
   async function deleteComment({
@@ -264,7 +282,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     );
   }
 
-  function addReaction({
+  async function addReaction({
     threadId,
     commentId,
     emoji,
@@ -273,7 +291,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     emoji: string;
   }) {
-    return fetchJson<CommentData>(
+    const reaction = await fetchJson<CommentUserReactionPlain>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}/reactions`,
@@ -285,9 +303,11 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
         body: JSON.stringify({ emoji }),
       }
     );
+
+    return convertToCommentUserReaction(reaction);
   }
 
-  function removeReaction({
+  async function removeReaction({
     threadId,
     commentId,
     emoji,
@@ -296,7 +316,7 @@ export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
     commentId: string;
     emoji: string;
   }) {
-    return fetchJson<CommentData>(
+    await fetchJson<CommentData>(
       `/threads/${encodeURIComponent(threadId)}/comments/${encodeURIComponent(
         commentId
       )}/reactions/${encodeURIComponent(emoji)}`,