+++ packages/liveblocks-node/src/__tests__/webhooks.test.ts
@@ -295,6 +295,31 @@ describe("WebhookHandler", () => {
       expect(event).toEqual(userEnteredBody);
     });
 
+    it("should throw if the rawBody is not a string", () => {
+      jest.useFakeTimers({
+        now: 1674850126000,
+      });
+
+      const webhookHandler = new WebhookHandler(secret);
+      expect(() =>
+        webhookHandler.verifyRequest({
+          headers: {
+            ...userEnteredHeaders,
+            "webhook-signature": generateSignatureWithSvix(
+              secret,
+              userEnteredHeaders["webhook-id"],
+              userEnteredHeaders["webhook-timestamp"],
+              rawUserEnteredBody
+            ),
+          },
+          // @ts-expect-error: we want to test invalid rawBody
+          rawBody: {},
+        })
+      ).toThrowError(
+        'Invalid rawBody field, must be a string, got "object" instead. It is likely that you need to JSON.stringify the body before passing it.'
+      );
+    });
+
     it("should throw if the signature is invalid", () => {
       jest.useFakeTimers({
         now: 1674850126000,