+++ packages/liveblocks-react/src/comments/CommentsRoom.ts
@@ -298,7 +298,7 @@ export function createCommentsRoom<TThreadMetadata extends BaseMetadata>(
 
         const threadId = createOptimisticId(THREAD_ID_PREFIX);
         const commentId = createOptimisticId(COMMENT_ID_PREFIX);
-        const now = new Date().toISOString();
+        const now = new Date();
 
         const newComment: CommentData = {
           id: commentId,
@@ -355,7 +355,7 @@ export function createCommentsRoom<TThreadMetadata extends BaseMetadata>(
         const threads = getThreads();
 
         const commentId = createOptimisticId(COMMENT_ID_PREFIX);
-        const now = new Date().toISOString();
+        const now = new Date();
 
         const comment: CommentData = {
           id: commentId,
@@ -417,7 +417,7 @@ export function createCommentsRoom<TThreadMetadata extends BaseMetadata>(
     const editComment = useCallback(
       ({ threadId, commentId, body }: EditCommentOptions): void => {
         const threads = getThreads();
-        const now = new Date().toISOString();
+        const now = new Date();
 
         const optimisticData = threads.map((thread) =>
           thread.id === threadId
@@ -468,7 +468,7 @@ export function createCommentsRoom<TThreadMetadata extends BaseMetadata>(
     const deleteComment = useCallback(
       ({ threadId, commentId }: DeleteCommentOptions): void => {
         const threads = getThreads();
-        const now = new Date().toISOString();
+        const now = new Date();
 
         const newThreads: ThreadData<TThreadMetadata>[] = [];
 
@@ -529,7 +529,7 @@ export function createCommentsRoom<TThreadMetadata extends BaseMetadata>(
     const createComment = useCallback(
       ({ threadId, commentId, emoji }: CommentReactionOptions): void => {
         const threads = getThreads();
-        const now = new Date().toISOString();
+        const now = new Date();
         const userId = getCurrentUserId();
 
         const optimisticData = threads.map((thread) =>